requirements.txt
Django==3.1.3
django-cors-headers==3.7.0
django-s3-storage==0.13.4
djangorestframework==3.12.2
postgres==3.0.0
zappa==0.52.0
